refactor : 로또 번호들을 LottoNumbers 일급 컬렉션으로 포장
