+++ README.md
@@ -16,6 +16,7 @@
 * 지난주 당첨 번호의 당첨 번호와 보너스볼이 중복되면 안된다.
 * 보너스볼, 로또 번호 원시값을 LottoNumber 객체로 포장한다.
 * 로또 번호, 보너스볼이 1이상 45이하의 숫자이다.
+* 로또 번호들을 LottoNumbers 일급 컬렉션으로 포장한다.
 * 당첨번호와 일치하는 숫자의 개수를 파악하는 기능
 * 로또 번호는 중복될 수 없다.
 * 로또 번호의 개수는 6자리이다.
