+++ src/test/java/domain/WinningLottoTest.java
@@ -3,6 +3,8 @@ package domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.util.Arrays;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class WinningLottoTest {
@@ -12,6 +14,7 @@ class WinningLottoTest {
     void generate() {
         // given
         Lotto lotto = new Lotto(new int[]{1, 2, 3, 4, 5, 6});
+
         int bonusBall = 7;
 
         // when
